+++ R/summarizeExperiments.R
@@ -15,5 +15,8 @@
 summarizeExperiments = function(ids = NULL, by = c("problem", "algorithm"), reg = getDefaultRegistry()) {
   assertExperimentRegistry(reg)
   assertCharacter(by, any.missing = FALSE, min.chars = 1L, min.len = 1L, unique = TRUE)
-  getJobPars(ids = ids, flatten = !setequal(by, c("problem", "algorithm")), reg = reg)[, list(.count = .N), by = by]
+  tab = getJobPars(ids = ids, reg = reg)
+  if (!setequal(by, c("problem", "algorithm")))
+    tab = flatten(tab)
+  tab[, list(.count = .N), by = by]
 }