+++ tests/testthat/test_addAlgorithm.R
@@ -15,11 +15,11 @@ test_that("addAlgorithm", {
 
   removeAlgorithms(reg = reg, "a1")
   expect_integer(reg$status$job.id, len = 2L)
-  expect_set_equal(levels(reg$defs$algorithm), "a2")
-  expect_set_equal(getAlgorithmIds(reg), "a2")
+  expect_set_equal(reg$algorithms, "a2")
+  expect_set_equal(reg$algorithms, "a2")
   expect_false(file.exists(getAlgorithmURI(reg, "a1")))
   expect_true(file.exists(getAlgorithmURI(reg, "a2")))
-  expect_set_equal(as.character(getJobPars(reg = reg)$algorithm), "a2")
+  expect_set_equal(getJobPars(reg = reg)$algorithm, "a2")
   checkTables(reg)
 })